app/requirements.txt
Django>=4.2,<5.1
wagtail>=6.2,<6.3
python-dotenv>=1.0.1,<2.0.0
openai>=1.40.5,<2.0.0
pydantic>=2.8.2,<3.0.0
django-htmx>=1.19.0,<2.0.0
gunicorn>=22.0.0,<23.0.0
dj-database-url>=2.2.0,<3.0.0
psycopg2-binary>=2.9.9,<3.0.0
boto3>=1.35.5,<2.0.0
django-storages>=1.14.4,<2.0.0
django-registration-redux
